Select features using `top_p` or `threshold`.

Feature selection using either the `top_p` or `threshold` features OR `cutoff` where cutoff refers to the absolute numeric value of the feature importance scores.

Usage

dual_filter(x, top_p, threshold, cutoff, maximize)

Arguments

x

a named numeric vector of scores per feature

top_p

an integer specifying the number of top-performing features to retain

threshold

a numeric with percentile of top-performing features to retain. For example, `threshold = 0.9` will only retain features that are in the top 90th percentile. A smaller value of threshold will select more features.

cutoff

a numeric with the value that represents the cutoff in the scores in `x` by which to retain/discard features.

maximize

logical to indicate whether `top_p`, `threshold` and `cutoff` are used to keep features where high scores = 'best' (maximize = TRUE) or where low scores = 'best' (maximize = FALSE).

Value

character vector of feature names to exclude

Details

`dual_filter` selects feature that are selected using either (`top_p`, `threshold`) or `cutoff` or both. If top_p/threshold and cutoff are both used then features are selected using OR. For example, if top_p selects features 1 & 2, and threshold selects features 1 & 3, then the selected features = 1,2,3.